+++ babel.config.json
@@ -1,0 +1,7 @@
+{
+  "presets": ["@babel/preset-flow"],
+  "ignore": [
+    "**/*.test.js",
+    "**/*.spec.js",
+  ]
+}
