+++ package.json
@@ -5,6 +5,7 @@
 	"homepage": "./",
 	"main": "src/electron",
 	"dependencies": {
+		"@babel/cli": "^7.8.4",
 		"@date-io/date-fns": "1.x",
 		"@fortawesome/fontawesome-svg-core": "^1.2.27",
 		"@fortawesome/free-solid-svg-icons": "^5.12.1",
@@ -65,6 +66,9 @@
 		]
 	},
 	"devDependencies": {
+		"@babel/core": "7.8.6",
+		"@babel/plugin-transform-flow-strip-types": "^7.8.3",
+		"@babel/preset-flow": "^7.8.3",
 		"@types/jest": "^25.1.3",
 		"electron-rebuild": "1.10.0",
 		"flow-bin": "^0.119.1",
