--- src/setup.js
@@ -1,2 +1,0 @@
-module.exports = async () => {
-};
